pyproject.toml
[tool.poetry]
name = "quick-borrow"
version = "0.1.0"
description = "System for self-checkout library"
authors = ["Štěpán Kotek <stepan.kotek@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
psycopg2 = "^2.9.7"
uuid = "^1.30"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-grappelli = "^3.0.7"
pytest-django = "^4.5.2"
isort = "^5.12.0"
black = "^23.9.0"
django-filter = "^23.2"
pyproject-flake8 = "^6.0.0.post1"
flake8-django = "^1.4"
flake8-bugbear = "^23.7.10"


[tool.poetry.group.dev.dependencies]
django-silk = "^5.0.3"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
known_django = "django"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
